Add Home and color palette screens
