+++ App.js
@@ -1,57 +1,21 @@
 import React from 'react';
-import { Text, SafeAreaView, StyleSheet, FlatList } from 'react-native';
 import { NavigationContainer } from '@react-navigation/native';
-import ColorBox from './components/ColorBox';
+import Home from './screens/Home';
+import ColorPalette from './screens/ColorPalette';
+import { createStackNavigator } from '@react-navigation/stack';
 
-const COLORS = [
-  { colorName: 'Base03', hexCode: '#002b36' },
-  { colorName: 'Base02', hexCode: '#073642' },
-  { colorName: 'Base01', hexCode: '#586e75' },
-  { colorName: 'Base00', hexCode: '#657b83' },
-  { colorName: 'Base0', hexCode: '#839496' },
-  { colorName: 'Base1', hexCode: '#93a1a1' },
-  { colorName: 'Base2', hexCode: '#eee8d5' },
-  { colorName: 'Base3', hexCode: '#fdf6e3' },
-  { colorName: 'Yellow', hexCode: '#b58900' },
-  { colorName: 'Orange', hexCode: '#cb4b16' },
-  { colorName: 'Red', hexCode: '#dc322f' },
-  { colorName: 'Magenta', hexCode: '#d33682' },
-  { colorName: 'Violet', hexCode: '#6c71c4' },
-  { colorName: 'Blue', hexCode: '#268bd2' },
-  { colorName: 'Cyan', hexCode: '#2aa198' },
-  { colorName: 'Green', hexCode: '#859900' },
-];
+const Stack = createStackNavigator();
 
 const App = () => {
   return (
     <NavigationContainer>
-      <SafeAreaView>
-        <FlatList
-          style={styles.container}
-          data={COLORS}
-          keyExtractor={(item) => item.colorName}
-          renderItem={({ item }) => (
-            <ColorBox colorName={item.colorName} hexCode={item.hexCode} />
-          )}
-          ListHeaderComponent={<Text style={styles.header}>Solarized</Text>}
-        />
-      </SafeAreaView>
+      <Stack.Navigator initialRouteName="Home">
+        <Stack.Screen name="Home" component={Home} />
+        <Stack.Screen name="ColorPalette" component={ColorPalette} />
+      </Stack.Navigator>
     </NavigationContainer>
   );
 };
 
-const styles = StyleSheet.create({
-  container: {
-    paddingHorizontal: 10,
-    paddingTop: 50,
-    marginHorizontal: 10,
-  },
-  header: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: 'black',
-    marginBottom: 10,
-  },
-});
 
 export default App;
