+++ screens/Home.js
@@ -1,0 +1,12 @@
+import React from 'react';
+import { Text, View } from 'react-native';
+
+const Home = () => {
+  return (
+    <View>
+      <Text>Hello World!</Text>
+    </View>
+  );
+};
+
+export default Home;
